Fix build error when compare int with size_t

+++ include/multigrid_pclomp/multi_voxel_grid_covariance_omp.h
@@ -351,7 +351,7 @@ protected:
 
   // Wait for all running threads to finish
   inline void sync() {
-    for(int i = 0; i < thread_futs_.size(); ++i) {
+    for(size_t i = 0; i < thread_futs_.size(); ++i) {
       if(thread_futs_[i].valid()) {
         thread_futs_[i].wait();
       }
